Trait async_macros::utils::future::IntoFuture

pub trait IntoFuture where
    <Self::Future as Future>::Output == Self::Output, {
    type Output;
    type Future: Future;
    fn into_future(self) -> Self::Future;
}

🔬 This is a nightly-only experimental API. (into_future)

Conversion into a Future.

Associated Types

type Output

🔬 This is a nightly-only experimental API. (into_future)

The output that the future will produce on completion.

type Future: Future

🔬 This is a nightly-only experimental API. (into_future)

Which kind of future are we turning this into?

Loading content...
Required methods

fn into_future(self) -> Self::Future

🔬 This is a nightly-only experimental API. (into_future)

Creates a future from a value.

Loading content...